fix: Fixed the bug that the text drawer could not display multiple texts at the same time
To do this, we're going to split the rendering into two phases. In the first stage, we prepare data for drawing that can be done immediately and brushes that need to be drawn uniformly. The second stage is the actual painting of the brushes that need to be prepared first