fix(image): correctly render BGRA texture formats
BGRA formatted images were being displayed with inverted red and blue
channels.

This commit introduces a check for BGRA texture formats and passes a
uniform to the shader. The shader then swizzles the color components
(from .bgra to .rgba) to ensure the image is rendered with the correct
colors.fix(image): correctly render BGRA texture formats

BGRA formatted images were being displayed with inverted red and blue
channels.

This commit introduces a check for BGRA texture formats and passes a
uniform to the shader. The shader then swizzles the color components
(from .bgra to .rgba) to ensure the image is rendered with the correct
colors.